+++ lizard_progress/tests/test_models.py
@@ -97,6 +97,7 @@ class LocationF(factory.DjangoModelFactory):
     """Factory for Location models."""
     class Meta:
         model = models.Location
+        django_get_or_create = ('location_code', 'activity')
 
     location_code = "SOME_ID"
     activity = factory.SubFactory(ActivityF)
@@ -160,6 +161,14 @@ class UploadLogF(factory.DjangoModelFactory):
     when = factory.LazyAttribute(lambda a: datetime.datetime.now())
     filename = 'test.txt'
     num_measurements = 1
+
+
+class ExpectedAttachmentF(factory.DjangoModelFactory):
+    class Meta:
+        model = models.ExpectedAttachment
+
+    filename = '1.jpg'
+    uploaded = False
 
 
 class TestUser(FixturesTestCase):
@@ -396,6 +405,69 @@ class TestMeasurement(FixturesTestCase):
         point = Point(0, 0)
         measurement.record_location(point)
         self.assertEquals(location.the_geom, point)
+
+    def test_setup_expected_attachments_deletes_missing(self):
+        measurement = MeasurementF.create()
+        expected_attachment = ExpectedAttachmentF.create()
+        expected_attachment.measurements.add(measurement)
+
+        # Sanity check
+        self.assertEquals(measurement.expected_attachments.count(), 1)
+
+        measurement.setup_expected_attachments([])
+
+        # Expected attachment was deleted
+        self.assertFalse(measurement.expected_attachments.exists())
+        self.assertRaises(
+            models.ExpectedAttachment.DoesNotExist,
+            lambda: models.ExpectedAttachment.objects.get(
+                pk=expected_attachment.id))
+
+    def test_setup_expected_attachments_adds_new(self):
+        measurement = MeasurementF.create()
+        measurement.setup_expected_attachments(['1.jpg', '2.jpg'])
+
+        self.assertEquals(measurement.expected_attachments.count(), 2)
+
+    def test_setup_expected_attachments_connects_to_existing(self):
+        # We have two measurements in the same activity
+        activity = ActivityF()
+        location1 = LocationF(activity=activity, location_code='CODE1')
+        location2 = LocationF(activity=activity, location_code='CODE2')
+        measurement1 = MeasurementF.create(location=location1)
+        measurement2 = MeasurementF.create(location=location2)
+
+        # Measurement 1 expects 1.jpg
+        measurement1.setup_expected_attachments(['1.jpg'])
+        # Measurement 2 also expects 1.jpg
+        measurement2.setup_expected_attachments(['1.jpg'])
+
+        # They are attached to the same expected attachment
+        self.assertEquals(measurement1.expected_attachments.count(), 1)
+        self.assertEquals(measurement2.expected_attachments.count(), 1)
+        self.assertEquals(
+            measurement1.expected_attachments.get().id,
+            measurement2.expected_attachments.get().id)
+
+    def test_setup_expected_attachments_raises_if_attachment_uploaded(self):
+        # We have two measurements in the same activity
+        activity = ActivityF()
+        location1 = LocationF(activity=activity, location_code='CODE1')
+        location2 = LocationF(activity=activity, location_code='CODE2')
+        measurement1 = MeasurementF.create(location=location1)
+        measurement2 = MeasurementF.create(location=location2)
+
+        # Measurement 1 expects 1.jpg
+        measurement1.setup_expected_attachments(['1.jpg'])
+
+        # It is uploaded
+        attachment = measurement1.expected_attachments.all()[0]
+        attachment.register_uploading()
+
+        # And now measurement 2 also expects 1.jpg -- that is an error.
+        self.assertRaises(
+            models.AlreadyUploadedError,
+            lambda: measurement2.setup_expected_attachments(['1.jpg']))
 
 
 @attr('slow')
@@ -652,3 +724,61 @@ class TestUploadLog(FixturesTestCase):
 
         self.assertTrue(latest)
         self.assertEquals(latest[0].when, date2)
+
+
+class TestExpectedAttachment(FixturesTestCase):
+    def test_detach(self):
+        """Test that expected attachments can be detached from a measurement,
+        and that if no measurements are left, the measurement is
+        deleted.
+
+        """
+        # Setup with two measurements
+        attachment = ExpectedAttachmentF.create()
+        measurement1 = MeasurementF.create()
+        measurement2 = MeasurementF.create()
+        attachment.measurements.add(measurement1)
+        attachment.measurements.add(measurement2)
+
+        self.assertEquals(attachment.measurements.count(), 2)
+        attachment.detach(measurement1)
+        self.assertEquals(attachment.measurements.count(), 1)
+
+        # Check that this still works
+        models.ExpectedAttachment.objects.get(pk=attachment.id)
+
+        attachment.detach(measurement2)
+
+        # Check that attachment has been deleted
+        self.assertRaises(
+            models.ExpectedAttachment.DoesNotExist,
+            lambda: models.ExpectedAttachment.objects.get(pk=attachment.id))
+
+    def test_register_uploading_creates_measurements(self):
+        # We have two measurements in the same activity
+        activity = ActivityF.create()
+        location1 = LocationF.create(activity=activity, location_code='CODE1')
+        location2 = LocationF.create(activity=activity, location_code='CODE2')
+        measurement1 = MeasurementF.create(location=location1)
+        measurement2 = MeasurementF.create(location=location2)
+
+        # They both expect an attachment named '1.mpg'
+        measurement1.setup_expected_attachments(['1.mpg'])
+        measurement2.setup_expected_attachments(['1.mpg'])
+
+        # The attachment
+        attachment = measurement1.expected_attachments.all()[0]
+
+        # Now it is uploaded
+        new_measurements = attachment.register_uploading()
+
+        # Both locations have a measurement added with the right parent
+        new_measurement1 = models.Measurement.objects.get(
+            location=location1, parent=measurement1)
+        new_measurement2 = models.Measurement.objects.get(
+            location=location2, parent=measurement2)
+
+        # And those were the sames ones as returned by register_uploading
+        self.assertEquals(
+            sorted([m.id for m in new_measurements]),
+            sorted([new_measurement1.id, new_measurement2.id]))